fix : refactor the search use case to use the combined search repository
