--- domain/src/main/java/com/cairosquad/domain/search/repository/SearchHistoryRepository.kt
@@ -1,14 +1,0 @@
-package com.cairosquad.domain.search.repository
-
-
-interface SearchHistoryRepository {
-    suspend fun getAllHistory(): List<String>
-
-    suspend fun getAllHistoryByQuery(query: String): List<String>
-
-    suspend fun clearAll()
-
-    suspend fun removeQuery(query: String)
-
-    suspend fun addQuery(query: String)
-}
