+++ domain/src/main/java/com/cairosquad/domain/usecase/search/SearchUseCase.kt
@@ -1,30 +1,28 @@
-package com.cairosquad.domain.search.usecase
+package com.cairosquad.domain.usecase.search
 
-import com.cairosquad.domain.search.repository.SearchHistoryRepository
-import com.cairosquad.domain.search.repository.SearchRepository
+import com.cairosquad.domain.repository.SearchRepository
 import com.cairosquad.entity.Artist
 import com.cairosquad.entity.Movie
 import com.cairosquad.entity.Series
 
 class SearchUseCase(
     private val searchRepository: SearchRepository,
-    private val recentSearchRepository: SearchHistoryRepository
 ) {
     suspend fun getSeries(query: String, page: Int = 1): List<Series> {
         return searchRepository.getSeries(query).also {
-            recentSearchRepository.addQuery(query)
+            searchRepository.addQuery(query)
         }
     }
 
     suspend fun getMovies(query: String, page: Int = 1): List<Movie> {
         return searchRepository.getMovies(query).also {
-            recentSearchRepository.addQuery(query)
+            searchRepository.addQuery(query)
         }
     }
 
     suspend fun getArtists(query: String, page: Int = 1): List<Artist> {
         return searchRepository.getArtists(query).also {
-            recentSearchRepository.addQuery(query)
+            searchRepository.addQuery(query)
         }
     }
 }
