+++ domain/src/test/java/com/cairosquad/domain/usecase/SearchUseCaseTest.kt
@@ -1,8 +1,7 @@
 package com.cairosquad.domain.usecase
 
-import com.cairosquad.domain.search.repository.SearchHistoryRepository
-import com.cairosquad.domain.search.repository.SearchRepository
-import com.cairosquad.domain.search.usecase.SearchUseCase
+import com.cairosquad.domain.repository.SearchRepository
+import com.cairosquad.domain.usecase.search.SearchUseCase
 import com.cairosquad.entity.Artist
 import com.cairosquad.entity.Movie
 import com.cairosquad.entity.Series
@@ -22,8 +21,7 @@ import org.junit.Test
 
 class SearchUseCaseTest {
 
-    private val searchRepository = mockk<SearchRepository>()
-    private val recentSearchRepository = mockk<SearchHistoryRepository>(relaxed = true)
+    private val searchRepository = mockk<SearchRepository>(relaxed = true)
     private lateinit var useCase: SearchUseCase
 
     private val dispatcher = StandardTestDispatcher()
@@ -31,7 +29,7 @@ class SearchUseCaseTest {
     @Before
     fun setUp() {
         Dispatchers.setMain(dispatcher)
-        useCase = SearchUseCase(searchRepository, recentSearchRepository)
+        useCase = SearchUseCase(searchRepository)
     }
 
     @After
@@ -49,12 +47,12 @@ class SearchUseCaseTest {
         )
 
         coEvery { searchRepository.getMovies(query) } returns movies
-        coEvery { recentSearchRepository.addQuery(query) } just runs
+        coEvery { searchRepository.addQuery(query) } just runs
 
         val result = useCase.getMovies(query)
 
         coVerify { searchRepository.getMovies(query) }
-        coVerify { recentSearchRepository.addQuery(query) }
+        coVerify { searchRepository.addQuery(query) }
     }
 
     @Test
@@ -66,12 +64,12 @@ class SearchUseCaseTest {
         )
 
         coEvery { searchRepository.getSeries(query) } returns series
-        coEvery { recentSearchRepository.addQuery(query) } just runs
+        coEvery { searchRepository.addQuery(query) } just runs
 
         val result = useCase.getSeries(query)
 
         coVerify { searchRepository.getSeries(query) }
-        coVerify { recentSearchRepository.addQuery(query) }
+        coVerify { searchRepository.addQuery(query) }
     }
 
     @Test
@@ -85,11 +83,11 @@ class SearchUseCaseTest {
         )
 
         coEvery { searchRepository.getArtists(query) } returns artists
-        coEvery { recentSearchRepository.addQuery(query) } just runs
+        coEvery { searchRepository.addQuery(query) } just runs
 
         val result = useCase.getArtists(query)
 
         coVerify { searchRepository.getArtists(query) }
-        coVerify { recentSearchRepository.addQuery(query) }
+        coVerify { searchRepository.addQuery(query) }
     }
 }
